experimental/opencog-cognitive-synergy/requirements.txt
# Bot Framework dependencies
botbuilder-integration-aiohttp>=4.15.0
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0

# OpenCog-style cognitive architecture dependencies
numpy>=1.21.0
scipy>=1.7.0
networkx>=2.6

# Additional AI/ML libraries for cognitive processing
scikit-learn>=1.0.0
pandas>=1.3.0

# Async and concurrency support
aiofiles>=0.8.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0